Rename constrained search methods to sqp
