+++ proj2/sqp_methods.py
